Added contour search
I went based off of the linked sites by Levin.
It seems to work fine on the symbols but there are issues when I used a picture from my old dataset.
We will need to check how much noise it can resist.

Because of formatting it seems it needs jpg as a type and I am not sure how to fix it...