added code for solving the equation and formatting
I added a few lines to trim operations that cause syntax issues

I also added code for formatting and evaluating the equations.
Since # is not used as a symbol for us I use it is as the placeholder for cdot.
For evaluating it gets turned into "*" and for latex it turns into \cdot